+++ GoalMogul-iOS/src/Registration/AddProfilePic.js
@@ -22,7 +22,7 @@ class AddProfilePic extends Component {
   render() {
     return (
       <View style={Styles.containerStyle}>
-        <Header name='John Doe' />
+        <Header name='John Doe' type='profile' />
         <View style={Styles.bodyContainerStyle}>
           <Text style={Styles.titleTextStyle}>Add a picture</Text>
           <View style={{ alignSelf: 'center' }}>